+++ Tpcc/Deferred.cs
@@ -28,10 +28,10 @@ namespace Tpcc
 			var s = db.ExecuteQuery("select NO_O_ID from NEW_ORDER where NO_W_ID="+wid+" and NO_D_ID="+did);
 			if (s.items.Count==0)
 				return false;
-			oid = (int)(long)s[0][0];
+			oid = (int)s[0][0];
 			db.ExecuteNonQuery("delete NEW_ORDER where NO_W_ID="+wid+" and NO_D_ID="+did+" and NO_O_ID="+oid);
 			s = db.ExecuteQuery("select O_C_ID from ORDER where O_W_ID="+wid+" and O_D_ID="+did+" and O_ID="+oid);
-		    ocid = (int)(long)s[0][0];
+		    ocid = (int)s[0][0];
 			db.ExecuteNonQuery("update ORDER where O_W_ID="+wid+" and O_D_ID="+did+" and O_ID="+oid + " set O_CARRIER_ID = "+carid);
 			db.ExecuteNonQuery("update ORDER_LINE  where OL_W_ID="+wid+" and OL_D_ID="+did+" and OL_O_ID="+oid+ " set OL_DELIVERY_DATE=date'" + DateTime.Now.ToString("yyyy-MM-dd") + "'");
 			decimal amount = 0.0M;
@@ -59,7 +59,7 @@ namespace Tpcc
             {
                 var s = db.ExecuteQuery("select DL_CARRIER_ID from DELIVERY where DL_W_ID=" + wid + " and DL_DONE is null order by DL_ID");
                 for (var i = 0; i < s.items.Count; i++)
-                    al.Add((int)(long)s.items[i].fields[0].Value);
+                    al.Add((int)s.items[i].fields[0].Value);
                 foreach (int k in al)
                     Carrier(k);
                 Thread.Sleep(30000); // 30 sec
